Fix mulitprocessing

+++ pysrc/cecog/errorcorrection/report.py
@@ -335,7 +335,7 @@ class HmmReport(object):
 
         for i, label in enumerate(track):
             name = self.classdef.class_names[label]
-            color = hex2rgb(self.classdef.hexcolors[name], mpl=False)
+            color = np.int(hex2rgb(self.classdef.hexcolors[name], mpl=False))
             image[size[1]-msize:size[1], i*size[0]:i*size[0]+msize] = color
         return image
